Remove run_nstate.py interface
run_nstate.py was created to reduce the number of states calculated
along a trajectory based on some criteria. Delete it now because it
already diverged from the main run.py interface and it was always
intented to be a temporary patch for a single set of trajectories.

It might be a useful option in the future as well, but in that case it
should be implemented as an option of the main surface hopping code and
not of the interface.