Fix lz_prob_conv option and update input manual
If statements containing the lz_prob_conv option were reversed after
a previous change to the keyword/procedure. This bugfix allows the
option to be used again.

Also activated the lz_min_dt option and updated the input manual to
explain both these options and the basics of how the QM interfaces
function.